requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
